remove unnecessary ui package and move UI class directly to the root directory of Duke package
